wic/engine: Fix missing parted autobuilder failures
OE-Core rev: 2a9147375f990d447a01b214c5a7d2fbd22e56cf broke the ability
to find parted as it may be in sbin which is not in PATH for some users
on some distros.

Iterate on the original patch to fix this and also fix the original problem.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>