Revert "wic: Set a miniumum FAT16 volume size."
This reverts commit 973663252ffa6f0a7eb62f34139242deb403a579.

FAT16 volumes do not have a minimum size of 8250 blocks
(== slightly over 4MB).  Exhibit A: floppy disks.

The original commit message suggests this was a workaround for a bug in
parted - in which case we should fix it there, or use the wic
`--fixed-size` option to pass down an enforced minimum from a more
context-aware point in the callstack.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>