remove 'create' and decouple from bitbake
The most common use for the wic command is to create images. But wic is a
front for several sub-tools and functions, only one of which is the ability
to create images. wic also contains very innovative ways of manipulating an
image, post-creation, without the need for elevated privileges. Currently
it only works with ext2/3/4 and vfat formatted partitions inside an image,
but these two are very commonly used as the partition formats in images.

One of the draw-backs of wic is that it is closely tied into bitbake; so
close, in fact, that it is not able to run unless bitbake is available.
This is a shame since its powerful ability to manipulate images
(post-build) is only available where bitbake is available. This means wic
can not be included in an SDK, it needs to be included in a (considerably
larger) eSDK in order to function.

The purpose of this branch and commit is to demonstrate the non-create
abilities of wic to stand on their own, outside of bitbake.

As it turns out, only the create and list sub-commands of wic actually
require bitbake. The remaining (i.e. the post-build image manipulation)
commands (ls, rm, cp, write, help) do not need bitbake's help.

This branch/version of wic shows it as its own stand-alone python package,
able to manipulate images, without having to be tied to bitbake, suitable for
inclusion in an SDK or on its own for manipulating images, post-build.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>